Convert entities to use doubles instead of floats.

+++ Src/Mojio/Events/Event.cs
@@ -86,7 +86,7 @@ namespace Mojio.Events
         /// <summary>
         /// Battery Voltage
         /// </summary>
-        public float? BatteryVoltage { get; set; }
+        public double? BatteryVoltage { get; set; }
 
         /// <summary>Creates a new object that is a copy of the current instance.</summary>
         /// <returns>A new object that is a copy of this instance.</returns>
